requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coverage==4.5.1
cryptography==2.2.2
Django==2.0.5
idna==2.6
pycparser==2.18
PyJWT==1.6.3
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22
